+++ packages/commonwealth/client/styles/components/component_kit/cw_divider.scss
@@ -3,6 +3,5 @@
 .Divider {
   background-color: $neutral-300;
   height: 1px;
-  margin: 16px 0;
   width: 100%;
 }
