+++ packages/commonwealth/client/styles/pages/snapshot/snapshot_proposal_content.scss
@@ -3,6 +3,7 @@
 .SnapshotProposalContent {
   display: flex;
   flex-direction: column;
+  gap: 16px;
   margin-right: 40px;
   max-width: 768px;
   width: 100%;
