+++ packages/commonwealth/client/styles/pages/view_proposal/proposal_header.scss
@@ -1,0 +1,131 @@
+@import 'shared';
+@import './index.scss';
+
+.ProposalHeader {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+
+  &.proposal-snapshot {
+    .snapshot-proposal-top {
+      display: flex;
+      border-bottom: 1px solid $neutral-200;
+    }
+
+    .proposal-content {
+      padding: 25px 0;
+    }
+  }
+
+  .proposal-top {
+    display: flex;
+  }
+
+  .proposal-top-left {
+    flex: 1;
+
+    .cui-input {
+      width: 70%;
+      padding-bottom: 10px;
+    }
+
+    .proposal-body-link {
+      display: flex;
+      flex-wrap: wrap;
+      gap: 8px;
+
+      > * div {
+        margin-right: 8px;
+      }
+    }
+
+    .proposal-body-meta {
+      align-items: center;
+      display: flex;
+      flex-direction: row;
+      flex-wrap: wrap;
+      gap: 8px;
+      margin-bottom: 12px;
+
+      svg {
+        color: $neutral-500;
+        cursor: pointer;
+      }
+    }
+
+    .proposal-meta-top {
+      display: flex;
+      justify-content: space-between;
+      margin-bottom: 12px;
+      height: 35px;
+
+      button {
+        margin-left: 17px;
+        border-radius: 8px;
+        height: 39px;
+        font-size: 18px;
+        text-transform: uppercase;
+      }
+
+      .proposal-meta-top-left {
+        display: flex;
+        align-items: center;
+      }
+
+      .proposal-meta-top-right {
+        display: flex;
+      }
+
+      .ProposalHeaderOnchainId {
+        font-size: 18px;
+        color: $black;
+      }
+    }
+  }
+
+  .proposal-content {
+    display: flex;
+  }
+
+  .proposal-content-right {
+    flex: 1;
+    overflow-x: auto;
+    margin-left: $discussion-comment-padding;
+    margin-bottom: 40px;
+
+    @include extraSmall {
+      margin-left: $discussion-comment-padding-mobile;
+    }
+
+    .proposal-content-meta {
+      padding-bottom: $discussion-meta-padding;
+      font-size: $discussion-meta-font-size;
+    }
+  }
+
+  .social-right {
+    margin-right: 10px;
+  }
+
+  .proposal-title {
+    margin-bottom: 6px;
+
+    .ProposalHeaderTitle {
+      display: inline;
+      font-size: $discussion-title-font-size;
+      font-weight: 600;
+      flex: 1;
+
+      .cui-tag {
+        font-weight: 400;
+        margin-left: 12px;
+        position: relative;
+        top: -2px;
+
+        .cui-icon {
+          margin-right: 3px;
+        }
+      }
+    }
+  }
+}
